requirements.txt
Django==1.6.5
South==1.0
dj-database-url==0.3.0
dj-static==0.0.6
django-bootstrap3==4.8.2
django-lockout==0.1.1
django-toolbelt==0.0.1
gunicorn==19.0.0
psycopg2==2.5.3
static3==0.5.1
wsgiref==0.1.2
